refactor(graph_view): extract pixel-spec/dim-windowing module (#80)
Move build_pixel_base_specs, dim_pixel_specs_window, dim_specs_window_core,
RowOids, force_dim_all_cells and apply_trace_dim out of graph_view/mod.rs into
a new graph_view/pixel_dim.rs (graph_pixels interop + per-frame dim windowing).
Pure code move: build_pixel_base_specs and dim_pixel_specs_window stay public
and are re-exported from mod.rs so external call sites are unchanged. The
tracing/folding pixel-dim test groups move alongside. is_base_update_row and
apply_merged_lane_dim stay in mod.rs (still used there) and widen to pub(super)
so the extracted core can call them.
